+++ packages/apiUtils/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @moralisweb3/api-utils
 
+## 2.3.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @moralisweb3/core@2.3.1
+
+## 2.3.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @moralisweb3/core@2.3.0
+
 ## 2.2.0
 
 ### Patch Changes